pyproject.toml
[tool.poetry]
name = "openai-swarm-ollama"
version = "0.1.0"
description = ""
authors = ["qiaochao <qiaochao@icourt.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.54.1"
python-dotenv = "^1.0.1"
swarm = {git = "https://github.com/openai/swarm.git"}
langchain-openai = "^0.2.6"
streamlit = "^1.39.0"
duckduckgo-search = "^6.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
